apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
neo4j==5.21.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==4.3.1
llama-index==0.10.68
llama-index-llms-openai==0.1.27
langgraph==0.2.16
pytest==8.3.2
sentence-transformers==2.7.0
python-jose[cryptography]==3.3.0
redis==5.0.8
prometheus-client==0.20.0
